Cleanup Weekday Indexing with Enum

+++ app/src/main/java/com/rikin/hydrohomie/features/hydration/domain/HydrationState.kt
@@ -1,8 +1,8 @@
 package com.rikin.hydrohomie.features.hydration.domain
 
 data class HydrationState(
-  val count: Float = 0F,
-  val goal: Float = 8F
+  val count: Double = 0.0,
+  val goal: Double = 8.0
 ) {
   val percent = count / goal
 }
